refactor(android): rename package names (#169)
