+++ packages/authentication/android/src/main/AndroidManifest.xml
@@ -1,3 +1,3 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="dev.robingenz.capacitorjs.plugins.firebase.authentication">
+    package="io.capawesome.capacitorjs.plugins.firebase.authentication">
 </manifest>
