--- packages/authentication/android/src/main/java/dev/robingenz/capacitorjs/plugins/firebase/authentication/GetIdTokenResultCallback.java
@@ -1,6 +1,0 @@
-package dev.robingenz.capacitorjs.plugins.firebase.authentication;
-
-public interface GetIdTokenResultCallback {
-    void success(String token);
-    void error(String message);
-}
