+++ packages/authentication/android/src/main/java/io/capawesome/capacitorjs/plugins/firebase/authentication/GetIdTokenResultCallback.java
@@ -1,0 +1,6 @@
+package io.capawesome.capacitorjs.plugins.firebase.authentication;
+
+public interface GetIdTokenResultCallback {
+    void success(String token);
+    void error(String message);
+}
