+++ packages/crashlytics/android/src/main/AndroidManifest.xml
@@ -1,3 +1,3 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="dev.robingenz.capacitorjs.plugins.firebase.crashlytics">
+    package="io.capawesome.capacitorjs.plugins.firebase.crashlytics">
 </manifest>
