+++ packages/messaging/android/src/main/AndroidManifest.xml
@@ -1,8 +1,8 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="dev.robingenz.capacitorjs.plugins.firebase.messaging">
+    package="io.capawesome.capacitorjs.plugins.firebase.messaging">
 
     <application>
-        <service android:name="dev.robingenz.capacitorjs.plugins.firebase.messaging.MessagingService" android:exported="false">
+        <service android:name="io.capawesome.capacitorjs.plugins.firebase.messaging.MessagingService" android:exported="false">
             <intent-filter>
                 <action android:name="com.google.firebase.MESSAGING_EVENT" />
             </intent-filter>
