--- packages/messaging/android/src/main/java/dev/robingenz/capacitorjs/plugins/firebase/messaging/GetTokenResultCallback.java
@@ -1,6 +1,0 @@
-package dev.robingenz.capacitorjs.plugins.firebase.messaging;
-
-public interface GetTokenResultCallback {
-    void success(String token);
-    void error(String message);
-}
