--- packages/messaging/android/src/main/java/dev/robingenz/capacitorjs/plugins/firebase/messaging/MessagingService.java
@@ -1,20 +1,0 @@
-package dev.robingenz.capacitorjs.plugins.firebase.messaging;
-
-import androidx.annotation.NonNull;
-import com.google.firebase.messaging.FirebaseMessagingService;
-import com.google.firebase.messaging.RemoteMessage;
-
-public class MessagingService extends FirebaseMessagingService {
-
-    @Override
-    public void onNewToken(@NonNull String token) {
-        super.onNewToken(token);
-        FirebaseMessagingPlugin.onNewToken(token);
-    }
-
-    @Override
-    public void onMessageReceived(@NonNull RemoteMessage remoteMessage) {
-        super.onMessageReceived(remoteMessage);
-        FirebaseMessagingPlugin.onMessageReceived(remoteMessage);
-    }
-}
