+++ packages/messaging/android/src/main/java/io/capawesome/capacitorjs/plugins/firebase/messaging/GetTokenResultCallback.java
@@ -1,0 +1,6 @@
+package io.capawesome.capacitorjs.plugins.firebase.messaging;
+
+public interface GetTokenResultCallback {
+    void success(String token);
+    void error(String message);
+}
